Move Minter/Wallet checker to wrappers
